Move safeCobraCommand interface into cliutil package and export it.
